refactor(collab): split collabSessionImport streaming/presentation
Move the fresh and incremental replay streamers into
collabImportStreaming and the source-derived timestamps plus display
refresh into collabImportPresentation; collabSessionImport keeps
importRemoteSession as the orchestrator and re-exports the options type.
The per-source serialization map stays with its only caller. No
behaviour change.
